refactor(config): reorganize into src/plugins/ structure
